refactor: `src/components` 하위의 컴포넌트를 `src/components/system`으로 이동한다 (#227)
