+++ src/components/auth/LoginForm.tsx
@@ -2,8 +2,8 @@ import type { LogInBody } from '@/lib/api/auth';
 
 import React from 'react';
 
-import Button from '@/components/Button';
-import LabelInput from '@/components/LabelInput';
+import Button from '@/components/system/Button';
+import LabelInput from '@/components/system/LabelInput';
 import { useFormError } from '@/hooks/useFormError';
 import { isEmail, isPassword } from '@/lib/validator';
